Instruction set is closer to J1s original, ...
The instruction set has been changed so it is closer to the original
instruction set of the J1 processor.

There are lots of changes in this commit, but there are still bugs that need to
be ironed out, specifically to do with RAM writes and reads. The assembler has
been changed to a simple test case that does not work as expected.

Other changes include improved filtering of signals when they are passed to
gtkwave.