Greatly improved h2.c, added back in j1 docs.
The 'h2.c' program has been greatly improved, but it is still a work in
progress. It will end up being an entire tool chain for the H2 CPU, an
assembler, a disassembler and a simulator for the H2 processor.

The disassembler can also be used to process the output of GTKwave, although
that has yet to be integrated into the build.

The current status of the program is:

Simulator: Still needs lots of work, can only process a subset of the
instructions available to the H2.

Assembler: The lexer has been added, and an outline of the grammar. The
assembler will be quite simple and should be used to process the output
of a proper compiler.

Disassembler: It works, although it is very simple. It should be suitable
for processing GTKwave output.

Most of the code in this commit is largely untested.

The J1 documents have been added back in.