Assembler and Disassembler have been improved.
Symbol files can not be loaded as well as saved, although they are not really
used for anything any the moment.

Code generation works with the assembler, but the grammar needs improving so it
looks more like Forth. Everything apart from the simulator is semi-usable. It
still needs a lot of testing however.

Some old code has been removed.