Reorganized project, h2.c improved, license, document, ...
The have been a lot of changes in this commit, lots of files have been renamed
and the project has been cleaned up. Some code has been reformatted.

The new assembler can now be used on the project, and it will be used for
testing the VHDL design from now on. The simulator is now mostly working as
well, apart from I/O functionality.

The disassembler can now lookup label and call addresses, if it finds a match,
then it prints the label name as well as the address.

A few bugs in various things have been fixed in h2 and it has generally been
improved upon.

The project is starting to be documented in "readme.md", but there is a copious
amount of documentation to write - which can be done when the project nears
completion and is in a somewhat workable state (like I have done for my
libforth project).