New instructions, generic stack size.
New instructions have been added to the "h2.vhd" CPU core, theses have been
added to the assembler as well. These instructions involved state and
operations that we already available to the CPU, but they were not being used.

Stack size has been made to be a generic in the core as well, any power of two
(within limits) can be used to set the stack size - if it becomes a problem.

Numeric input, error handling, and initialization have all been improved in the
assembler.