Improved LED segment interface, added [char].
[char] has been added to the assembler, and a bug has been fixed in it relating
to the generation of literals in certain circumstances.

The LED segment driver has been improved, the interface is more natural to its
function.

uart_top.vhd has been removed, an equivalent may be added back in later, one
that works.

There have been a lot of whitespace changes as well.