Added back in interrupts, ledseg, reorganization, ...
Interrupt handling has been changed so the logic for it is in "irqh.vhd", it is
very primitive at the moment and completely untested but should not interfere
with the CPU in normal running. The interrupt signals lines have also been
added back in.

Block RAM now has a write enable.

The "util.vhd" has been reorganized, some entities have been taken out of it
and moved into other files.

The LED segment display driver has been simplified and each segment can now be
addressed individually.

The input buttons have been debounced, but not the switches as of yet.