LED segment driver, interrupts, warnings, ...
The LED segment driver now works correctly, it needs its update rate
increasing. Its interface has now changed, values are input in BCD with a few
extra characters.

Comments changed from "--!" to "--|" are responsible for most of the changes in
this commit.

The project now synthesizes with only one warning, there are still warnings in
the bitfile generation.

Calls do not seem to work at the moment, but the rest of the h2.vhd CPU has not
been tested. This might be due to incorrect code generation or because of the
CPU.

Some generic functions have been added to the project, they should be used in
more places to simplify signal definitions.

Some buttons have been added to the list of interrupts that are available -
these are place holders until something more useful is found for them.