Only minor changes, mostly relating to coding style.
Most of these changes are related to coding standards.

The LED segment has been improved (yet again), so it uses a generic timer that
can wait in a time specified in micro seconds instead of the user of the module
having to specify a counter size. This new way of debouncing things should be
used throughout instead.

An outline of how I/O should be handled in the C simulator has been added as
well.