Added UART test bench. Improved h2.c simulator.
A test bench for the UART has been added and code for exercising the UART
implemented. The UART is the next thing that will need working on so a
bootloader can be implemented. Currently something along the chain from the
UART to the h2 core is dropping characters.

The "h2.c" simulator has been improved. There is still a lot of work to do
until it can simulate the H2 System on a Chip with any kind of accuracy, but it
would be useful for developing on before that.