Interrupt handler masks, switch debouncing, ...
Apart from a few minor changes the main things in this commit are:

Interrupt handling now has masks which can be applied to the interrupts,
allowing them or disabling them. This functionality has not currently been
tested (none of the interrupt handling has been tested in a while).

The input switches are now fully debounced before being passed to the rest of
the system, like the buttons are.