Added more utils with test bench, new docs
The utils file has been improved by adding an adder and a half adder. This
includes a test bench, there is still more to add.

The documents have been updated, the word document has been removed and an RTF
file has been added. This really should be converted to a markdown document
with links to the embedded images. The PDF file has been regenerated as well.