Improved eForth docs and utility functions.
The original eForth implementation has been added which will be used as a guide
to implementing eForth on the H2. The documentation for it has been translated
to markdown as well, from this a PDF and HTML can then be generated.

The functions in "util.vhd" have been expanded upon and improved and a test
bench for the functions within added.