Restructuring project to use records.
The project is being restructured to use records instead of large cumbersome
lists of signals. This does place limitations on the types used in the
interfaces due to the limitations of VHDL - which are largely solved in
VHDL-2008, alas it is not available in most simulators or synthesis tools.

The changes have been inspired by the "VHDL two process method", see
"www.gaisler.com/doc/vhdl2proc.pdf", although a lot of the actual logic would
have to be rewritten to effect all the methods it advocates. Changing the
interfaces to components is the first start.