Adding test code for FIFO version of UART.
This commit adds test code for a UART buffered with a FIFO, it currently does
not work but that could be for many different reasons from the VHDL
implementation to the FORTH code. It has not been proved to work in simulation
either, but this commit should contain most of the changes in the final
product. This commit also removes the instantiation of the UART which does not
use a FIFO. Hopefully this will remove some of the problems with the UART - but
it probably will not!