Updating documentation and register map.
Minor modifications to the register map are being made, this is being done in
tandem with updates to the documentation in "readme.md" which will now detail
the operation of the CPU and its peripherals, which have not been described so
far.

The "hack.c" program has been renamed to "uart.c", it is no longer a hack and
most of its functionality is not needed.

The PS2 register has been simplified, it now occupies only one register.