Improved register map documentation, ...
The register map documentation has been improved, there is still a lot of
documentation to write before it is complete, but it is a useful reference.

Some of the generic options have been changed in the project for "top.vhd". The
number of interrupts is no longer changeable as setting it to a value other
than 8 would cause the project to not compile as the interrupt routing would
have to change. The UART FIFO depth has been added as a generic option however,
as this can be changed.