Character alignment, documentation, ISR, ...
The major change in this commit is the fact that RAM access with loads and
stores (but not jumps and IO loads and stores) now discard the lowest bit in
the address so it can be used by applications for string operations.

The documentation in "readme.md" has been improved a lot - but there is still a
lot to do, the final documentation will have to be about 4 times larger to be
useful and detail everything. Pictures will need to be added.

The interrupt service routines have been updated and they come from different
sources now, which has been documented.