Added counted strings, fixed simulator, ...
The simulator has had some bugs fixed relating to the order in which operations
take place.

Counted strings have been added to the syntax of the assembler, the test code
does not quite work and there are differences still between the simulator and
the target board in how the code behaves.

The "uart.c" program has been modified, it still needs modification so it can
be used interactively.