Adding new Forth words, compiling Forth headers
At the moment the development is being done in the simulator as it appears not
to be working correctly in the hardware. More Forth words have been added in
"h2.fth" and a dictionary is starting to take form. "words" has been added, but
it needs access to the previous word pointer, access to which will have to be
added in "h2.c".