GUI: Switches, scaling, LEDs. Added CPU util comments
The GUI simulator has been improved a lot, scaling now works which is needed
for collision detection. The switches now look as if they have two states.

The GUI simulator could be improved with font scaling, currently a bit mapped
font is used, which does not scale at all.

The 8 segment display driver has been added as well.

Some comments about adding a tiny CPU to the "util.vhd" have been added, this
would be a hundred line CPU, incredibly simple and designed to be as small as
possible. It would act more like a programmable state machine than an actual
CPU with a very small address space.