Integrate H2 simulator with GUI program
The GUI simulator can now be used with the H2 core. It would need more work to
be fully functional, but it is usable and simulates more of the peripherals. It
would be better if the H2 core took its input from GUI program and not the
command line (for many reasons), and here are many outstanding to-do comments.