Removing threads from GUI program.
Threading is to be removed from the GUI program, which means locks and blocking
input can be removed. This is done by using a FIFO in the keyboard handling
callback, it needs to be tested and information gathered to find out if
freeglut calls all of its callbacks from a single thread.