Moved FIFO code, added EXIT optimization
The FIFO code has been moved from "gui.c" to "h2.c", this is in anticipation of
using it for the optimizer and to use it to simulate the FIFOs in the H2 SoC.

EXIT optimization has been added, which can be optionally disabled. It merges
EXIT with the previous word definition if that is possible.