Added temporary fix for UART transmission.
A temporary fix for the UART transmission has been added until the problems
with the TX FIFO have been fixed in the VHDL, it appears the FIFO full signal
is being optimized out for some reason, most likely a bug in the control logic
in either "top.vhd" or "uart.vhd". The fix is to delay a millisecond after
transmitting a character. The RX FIFO appears to be working correctly.