Turned optimizations off, fixed minor bug in uart.vhd
Optimizations have been turned off for now, they seem to work in the C
simulation but not in the hardware.

A minor bug in the UART module has been fixed, however it seems there are still
problems. I think this can be resolved by changing the UART core to emit a busy
line and using that as a control signal for the TX FIFO.