VGA Terminal handling, GUI debug & bug fix
The VGA output routines now respond to carriage returns so it moves the
cursor on to the next line.

The GUI program has had a bug fix relating to button pressing, release now do
not have to match the target area to turn a push button off.

More debugging information is also displayed, which can optionally be
controlled by pressing the F12 key.