CPU-ID instruction added, h2.vhd start address
A few more comments have been added.

The major change is adding a CPU-ID instruction, the value currently is set to
"CAFE" in the VHDL H2 core and "DEAD" in the simulation. This can be used to
differentiate the simulation from the hardware. A mode in the simulator should
be added to make them equal.

For some reason, adding an extra instruction has sped things up a little
according to the timing report generated by the synthesis tools. I have no idea
why, minor changes like these can sometimes have an effect, unfortunately for
completely unknown reasons!