Added LFSR in, char printing, UART program, ...
The linear feedback shift register has been added as a hardware component, this
is mainly just for testing the hardware and not as a useful component for the
SoC. It has also been documented and added as a simulated component.

Character printing on hex dumps has been added to the debugger.

The UART program has been improved, slightly, it still needs work as characters
seem to be lost when transmitting from the board to the PC, I am not sure where
the loss is occurring.