Improved documentation, added more debug commands
Some new debug commands have been added to the CLI debugger: File saving,
pushing values and popping values has been added to the debugger.

The "readme.md" document can be turned into a PDF file or a HTML file with a
generic make target for markdown to PDF/HTML conversion.

A lot more documentation has been added, but there is a lot to do before it is
complete - not to mention actually finishing the Forth system!