documentation, makefile, symbol bug, moved files
The documentation has been improved with a GIF of the running simulator showing
it working, and a picture of the waveform in GTKWave. These are both referenced
in "readme.md". More text has also been added.

The external documentation has been moved to the "info/" directory.

The makefile should work on Windows, provided the correct toolchain and
libraries are installed, as well as under Linux.

A bug in the assembler has been fixed, the symbols for built in words were not
added to the symbol table if they are added to the dictionary with the correct
compiler directive.

The io_pins_tb has been commented out, but left in, with a note as to why - it
causes problems with GHDL 0.29 under Windows, but not under Linux.

Some new Forth words have been added.