Improved simulator with UART ouput, ...
There have been some tiny changes elsewhere in the system, for example
"number?" in "h2.fth" now only produces two arguments instead of 4!

The main differences if improvements to the GUI simulator, which now includes a
small terminal emulator for the UART output, instead of dumping its output to
standard out. It is quite slow however.