Fixed bug in GUI, basic word compiling, comments
Some comments have been added to "readme.md" about using Forth as a bootloader
and about Forth block editors.

A method for compiling words into the dictionary has been added, however it is
crude and is primarily for testing purposes.

A bug in the GUI code has been fixed, it was a length related one, the wrong
size was used when looping through an array of 8-Segment displays.

Some C code has been added to "h2.c" that deals with code optimizations and
compiling variable headers, it does not quite work correctly.