Semi-working Forth compiler on the simulator
A very fragile Forth compiler and interpreter has been made that works in the
simulator. It does not handle errors well.

This commit involves changes to the compiler in "h2.c" as well as some
additional Forth code. The optimizer has also been fixed and words can be
defined that the compiler itself can use (although the word headers had to be
defined as well, which might not needed). The optimized code, nor the new
interpreter loop, have not been tested on the target board yet.