Whitespace, fixed name packing, and more
A lot of the changes in this commit are simple whitespace fixes, these really
should go in their own commit, but I usually remember fix whitespace part way
doing something else.

Word definitions were failing for works with an even number of characters in
them, due to a failure to align up addresses, this has been fixed.

"rdepth" and "depth" have been renamed to "rp@" and "sp@".

The maximum depth of the stacks is now stored in the "h2_t" structure, which is
useful for debugging purposes.

Some unneeded words have been commented out to save on space.

"create" and "variable" have been added, they seem to work.

"scan" and "skip" have been refactored so as to use a common work to do the
work, "lookfor".

Some more comments have also been added.