Making changes related to return stack alignment
These changes are to bring the H2 core more in line with the original J1 core,
the program counter when it moves to the return stack and from the return stack
is converted to and from a character to a cell based address (pc*2 -> rstk,
rst/2 -> pc).

The changes do not quite work, there are still problems, but hopefully they
will be fixed soon. The changes will also need to be tested in hardware.