Interrupts now in simulation, h2.vhd interrupt fix
The interrupt code in "h2.vhd" has been fixed so the new return behavior should
work for interrupts.

Interrupts were not tested in the last commit, but the return to PC and vice
versa changes worked in the hardware in the previous commit.

Interrupts are now available in the simulation and some test code (that does
not seem to work properly) has been added to the code base.