Fixed minor bug in depth/rdepth, comment
Added comment to "pick", it does not work when the variable stack depth is big,
it mashes the return stack.

The "depth" and "rdepth" or "sp@" and "rp@" primitives have been changed
slightly in "h2.vhd", the top bits of the top of the stack were not been
cleared when those instructions were being run.