Fixed memory map, can use more space
The memory map has been fixed, but this seems to slow down the device according
to the synthesis report - it appears to function properly on the device
however. This is going on a branch in case it does not work out, but there
seems to be no problem for now.

The problem was down to using the wrong bits to control when ram reads and
write occur, the 14th and 13th bits were being used instead of the 15th and
14th.