Variable headers, location, token, ...
Variables can now be compiled with headers and a new location word exists for
making variables with no header. The variables can now be called from within
the Forth interpreter loop.

The documentation has been updated to reflect changes to the memory map, as
well as adding new bits to the grammar and to-do comments.