ANSI Escape Sequences, rationalize VGA color scheme
ANSI Terminal Escape Sequences have been added, but not tested in the hardware,
they are output on the UART only as the VGA terminal driver does not understand
the sequences (yet). The VGA color scheme has been rationalized so the same
color related words can be used for the VGA control register as well as ANSI
color codes.