Simulator now selects screen, ANSI words
The GUI simulator now displays the correct screen if the screen select bit is
set in the control register for the VGA display.

Some words for ANSI terminal codes have been added, they have been commented
out to save on space however, conditional compilation should probably be
added to the assembler so modules can be selected.