Adding an interface to the Nexys3 on board memory
This is an experimental interface which has not been tested in the hardware at
all, and it might be conceptually flawed - depending on how the memory
interface works (the documentation is not too clear). At the moment it will
interface with just the Non-Volatile flash, but could be extended to the
Volatile RAM on board.