Updated simulation, memory interface
The simulators have been updated to include a memory interface, but it is not
quite exactly how the hardware works.

The VHDL memory interface has been updated, it appears that I can read memory
but not write to it. I do not know why.

This memory interface branch is still very much a work in progress.