Fixed divider, with code from divider branch
This code has been tested in hardware on the divider branch with a modified top
level module that integrates a divider, this will not be included on the master
branch, but it proves the restoring divider works in the hardware. The module
could do with modifications however, it should have a better interface and
provide the remainder as an output as well.