Fixed memory issue? Added RAM CS
The memory issue seems to be fixed, although the non-volatile RAM seems to be
quite volatile, this might actually just it being overwritten when the device
is reprogrammed, it needs further testing. A value can be written to and read
back from the memory.

A chip select for the cellular RAM has been added added.