Updated documentation, block layer
The documentation has been updated, the new registers for the external memory
devices have been documented, although a lot more could be done to describe
what is going on.

A set of words for a block layer in Forth have been added, they seem to work in
the hardware and in the simulation.

Some comments in "h2.c" have been removed, they were duplicates of comments
that can be found elsewhere in the repository.

It appears that the SoC is writing to the volatile RAM, not the non-Volatile
RAM, due to a bug in "top.vhd", this has been commented. The Volatile RAM
interface works, the Non-Volatile one does not, although they are meant to have
a very similar interface.

The commented out words have been added back in for now.