More Forth words, found bug, memory interface
More Forth words have been added to the interpreter, and this has uncovered a
problem, too many bits are being used in the PWD field of a word to store
information about the word now that the address space has been increased.

The memory interface has been change slightly, but it still does not quite
work when it comes to the Non-Volatile memory.

Commented out code for a block editor has been added.