Improved 'see', generate "text.bin", GUI debug
The GUI program now allows the user to single step the execution of the core or
halt it. It also reads in the initial contents of the VGA memory from the
"text.bin" file.

The makefile now generates the "text.bin" file from scratch if it does not
already exists, it does this with "text.c".

"see" has been improved, it now detects if cannot decompile a call to a
function instead of giving incorrect results.