Interrupts work in simulation, login code, msg
The interrupts have been tested and work in simulation, it has yet to be seen
if the same is so in the hardware.

Some commented out code for a login and user management system has been added,
new users can be added and listed, and they can login - although only one at a
time. The code is commented out and not used by default, it is just an
interesting curiosity, it uses an insecure password hashing mechanism for
example amongst other oddities.

The word 'msg' has been added, the plan is to use this to recall error messages
stored in the non-volatile memory once that has been sorted out.