Added hack for TX FIFO full problem
A hack for the VHDL for the TX full FIFO has been added, so the TX FIFO full
line goes high whenever the TX UART is busy, and so data is not lost. Although this
does not allow the forth implementation in the hardware to fully utilize the TX FIFO
the simulation can without changing the software. The GUI simulation is
subsequently a lot faster.