GUI swapped switch order, data sheets, VGA, ...
In the GUI simulator the order in which LEDs and switches are drawn have been
swapped.

Data sheets have been added in an archive for the Nexys 3.

Experimental color bits have been added so each individual character can be
colored on the VGA display, it is a work in progress. Perhaps bits could be
used for other things.

Some of the Forth code for the memory interface has been changed, it still does
not work for the non-volatile memory, but seems to work for the SRAM - they
should have a very similar interface.