Added NVRAM word set
A word set for accessing the non-volatile PCM non-volatile memory has been
added, it is a work in progress and the complex memory controller will need to
be replicated in the simulator. The word set is mostly working. It will have to
be integrated into the Forth "block" word set.