block help, reorganize top.vhd, ...
There have been many minor changes, some new comments, but the major changes in
this commit are two fold: The I/O decoding has changed so that writes are not
in the same process as reads. This breaks things up (and seems to speed up the
design slightly). The second change is the addition of a program 'block.c' for
converting a specially formatted text file (with line limits of 64-characters)
into a binary file that could potentially be loaded onto the target board.
An example file will help text has been added as 'nvram.txt'

There are some new forth words as well, and more forth words have been hidden
from the user to save on space.

Reading from the VGA display memory has been fixed, but this needs more
testing.