Adding memory simulator for flash memory
A simulator for a flash chip with a Common Flash Memory Interface type memory
controller <https://en.wikipedia.org/wiki/Common_Flash_Memory_Interface> has
been added, it is not fully implemented, nor liable to be correct in every
circumstance, but it is useful for testing out Forth programs in simulation.

The word set for operating with memory has been changed a little as well.