GUI Terminal Emulator, more Forth words, ...
The terminal emulator in the GUI simulator has been improved, it now deals with
terminal escape codes, there is still more to do, especially relating to the
attribute bits, but a subset of the ANSI terminal escape codes have been
implemented.

Some more Forth words have been added, and some moved into the "nvram.txt", of
note is the word set related to ANSI escape codes, this is now available to the
interpreter so it can load it at run time if needed.