Adding VT100 Terminal Emulator
This is a work in progress, this should take the burden of terminal control off
the Forth core and put it into hardware by emulating a VT100 terminal. The
first step is to get the hardware working, then the simulations. Currently the
build is failing at the synthesis step and the simulation of the system has not
been tested.