Terminal interface partially working.
The simulations have been improved so that they are partially functional, at
least allowing characters to be drawn to the screen.

The hardware device is almost working, with basic functionality that can
process spaces, backspaces, carriage returns, line feeds, however there are
still a few problems and bugs. This is a good days work however, and it is
progressing quite well.