GUI memory simulation fix, fix parse, word factor
A bug in the memory simulation present in the GUI simulator has been fixed.

The "parse" word has also been fixed so it consumes the space after a word.
This causes problems for Forth words like '."' when given an empty string - but
that is normal for Forths that do their parsing with 'word'.

Some more words have been factored and changed around.

This first block is loaded off disk at startup.