Updating I/O map
The way I/O works has been updated so that the VT100 interface, consisting of a
P2/2 Keyboard as input and a text mode VGA controller, act as a UART. The Forth
code has been changed to reflect this.

The simulations have also been updated, although the GUI simulation has not
been tested.