Removed unneeded signals, Interactive VHDL sim, ...
Some example plotting Forth functions have been added, which will need more
thinking about. They use ANSI escape codes to position the plotting cursor and
expect and array of 80 cells, and a scaling factor for the height.

A method for interacting with the VHDL simulation has been added, but it needs
to be hooked up to the rest of the system, it reads from standard input (and
will write to standard output).  This will be connected to a simulated UART,
which will then be connected to the rest of the system.

Some unneeded signals have been removed.