106.17MHz/20B4 Cells: Bug Fixes/To dos
A problem where the parser/lexer leaked memory has been fixed.

The loading and saving of the block ram, whilst slower than it was, should not
be endian agnostic. The code for loading and saving this has been factored out
into a set of routines which are now used in both the GUI and CLI simulators.

'store' and 'nop' are now hidden instructions, there was no need to export them
for use by the user.

A bug with the terminal attributes being set has been fixed, the entire screen
resets to the default attributes when an ANSI reset command is processed.

D-Pad key presses have been added to the GUI simulation, they can now be
processed by the SoC, they also trigger interrupts on a state change if they
are enabled, this has been tested in the simulation.

'variable' and 'constant' declarations can now be hidden, which is now part of
the Pseudo-Forth compiler grammar in 'h2.c'.

A few bugs in 'recurse', and some other words, have been fixed, which is
related to how ';' now links in a definition.

Some code has been moved from 'h2.fth' to 'nvram.txt', and some of the
documentation has been improved.