106.04MHz/2080 Cells: Removed reset line
The reset line is now left unconnected to the D-Pad, the wait line is still
attached to the center button as otherwise an incredibly bad timing report is
generated (82MHz) - which is odd, but the system still seems to work.

The 'rx' signal is now also not put into the switches signal either. There was
no need for it.