103.00MHz/2078 Cells: Documentation, Interrupts, ...
Some more documentation in 'nvram.txt' has been added, as well as some more
comments.

The 'embed' submodule has been removed, this was causing problems with the
Travis CI system (due to an SSH URL being used for the submodule).

A missing target for the makefile has been added. The makefile still needs
improving, the 'synthesis implementation bitfile' targets do not depend on
each other, although the Xilinx toolchains make it a pain to do.

The way the interrupts work have been changed, and some problems identified.
Some of the interrupt inputs were level interrupts when they should have been
edge triggered. There is still some problems with the interrupt hardware, it
appears to be a little more stable (but still unusable). I believe this must be
solved within the H2 core instead, perhaps by saving/restoring state when the
interrupt comes in, or more state? Better simulations will need to be done.
Some interrupt related changes in the H2 core have slowed things done, they
maybe changed back later. The interrupts really need a lot more investigation.

A bug in the GUI VT100 simulator has been fixed, which caused the game 'YOU ARE
DEAD' to be unplayable. This was due to the way cursor placement was done.