103.00MHz/2076 Cells: Documentation, color
The documentation in 'nvram.txt' has been updated.

The main code for the 'embed' virtual machine has been added, this should be
ported to the H2 core and a new toolchain devised.

Colorization options have been added to the decompiler, which allow for color
to be displayed in GTKWave and on the command line for different instruction
types.