95.571MHz/2086: Interrupts "working" modulo bugs
The interrupts are now working...but only interrupts now are delayed
until a branch instruction occurs (and not a 0-branch, just a branch).
If a branch instruction does not occur, the interrupt never gets run.
The reason for this is because when an interrupt can interrupt any other
instruction (I suspect some ALU instructions might be alright, depending
on what flags are set) *something* goes wrong, causing problems,
corruption and system instability.

This branch has another problem; although the image seems to work
perfectly on the device, it now reports that maximum frequency the
solution can run at is 95.571 MHz, much lower than the 100MHz the system
is meant to be running at.

Until a proper fix can be implemented, this solution is to be kept on a
branch.