98.539MHz/2096: Interrupts work!
Interrupts work for all instructions now! However, the system is
*theoretically* too slow, according to the synthesis report. The fix
came about by changing how interrupts are handled entirely; the current
instruction to be executed is intercepted in a single place and not all
over the shop. I may be able to speed things up (and increase latency of
interrupts by a clock cycle) by inserting registers in the correct
places...