109.542MHz/1A28: Sped up system, smaller image
* There have been a few minor changes here and there, the simulator has
been updated to reflect the new ALU instruction, the CPU-IDs for some
things have changes, the makefile spits out timing information right
after synthesis, there have been a few comment related changes, the
test bench configuration has been modified so it is in verbose mode
(like it should be).
* The eForth image continues to shrink whilst retaining largely the same
functionality. The was mostly done by hiding/removing test words (now
that interrupts work fully) and refactoring some words.
* The biggest improvement is the speed up of the system. This was
accomplished by adding an ALU instruction which performs what 'literal'
does, and changing some of the control logic elsewhere, so there are
fewer levels of logic in between the ALU and what controls it. The
simulator could be updated so it reflects the hardware more accurately,
but it is not necessary.

Note: The maximum return stack usage of the system is 32 deep, this is
too high! Although the stacks are 64 deep it would be best to use under
32 and not 32. This is an improvement of the 35/36 it was recently, but
something has knocked it back up to 32 from 30, and I forget what. This
should be investigated.