108.958MHz/1A1E: Breaking FIFO/UART test changes
This is the start of a test branch for new FIFO/UART changes which
should (eventually) improve some things and how the UART buffers its
TX/RX data. The UART does not work correctly yet! This commit breaks the
UART!

* A new FIFO has been implemented in the 'util.vhd' file. The old one
remains in there commented out.
* The way the 'stop', or wait state, in the H2 CPU works has changed, it
works in the same was as interrupts in that it replaces the current
instruction to be executed instead of by lots of little bits of logic
around the CPU.
* The UART has has some of its high level control logic reworked.