109.974MHz/1A6E: Improved 'see', ien!/ien?, ...
* In the eForth image the decompiler, 'see', has been improved a lot,
the code is clearer and more extensible.
* The 'ien!' and 'ien?' instructions have been improved, both could now
be extended as a more generic CPU status instruction. 'ien!' now sets T
to N so a -1 can be issued, making the instruction less awkward to use,
and 'ien?' does not return a boolean but a packed bit-field (with only
one status bit at the moment).
* Some of the RAM BUS interface signals have been renamed. A lot of
signals should be renamed really so they abide by a stricter naming
regime and coding standard.
* The compiler and simulator has been updated to reflect the instruction
changes.
* Comments have been removed/updated/added