109.974MHz/1A6E: embed meta-compiler working
The embed meta-compiler is now producing working images. There is more
to do in order to support all the features of the Nexys-3 and there are
still some problems with the new meta-compiled image.

Problems include:
* The interrupt vector table is not present, in-fact it has junk in it
* Switching on the CRC check causes the image not to boot - sometimes.
It must be failing the check, or there is some other problem.
* Error handling is not working correctly, the 'embed' VM allowed you to
set the stack address the stack pointers directly, You cannot do this in
the H2. On errors, these stack pointers were manipulated directly.
Something goes wrong with this.
* Documentation has been removed from the 'embed.fth' file, this should
be added back in.
* The 'embed' VM is present, and it is very similar to the H2 processor,
as it is derived from the simulator for it. It would be good to make the
embed VM work like the H2 again, and then get the meta-compiler working
on the H2 for a self hosting system.