109.974MHz/1A6E: Async/Sync reset, delay, h2 rst, ...
There are many changes in this commit, although they are all systematic
changes, the include the addition of a generic that can be used to
select between asynchronous and synchronous resets in each module, and
another generic for simulating delays introduced by gates (which is
ignored for synthesis purposes).

A module for generating a reset at startup has also been added, which
means there is one fewer warning in the project, and a minor hiccup with
the reset at startup has been fixed in the h2 core, where it would
continue to execute instructions if a reset was occurring.

There appears to be a bug in SoC, if you hammer the UART at startup
before the H2 core has had time to do anything, it becomes unresponsive
and/or gets weird errors relating to stack overflow.

The 'embed.fth' image is now in a semi-working state, the I/O is no
longer working correctly (or was never checked in a working state).