110.643MHz/1A6E: Miscellaneous changes
* Added more assertions to the H2 core, as well as comments about them.
It would be a good idea to assert as much as is possible so various
properties of the H2 core can be guaranteed or debugged on failure.
* Minor formatting changes in the VHDL code
* Some more comments about the direction of the project and things to
do, such as potential new modules and the like.
* Improved the jitter mechanism for the clock generation module and for
the test-bench.
* Minor VGA changes. Still looking buggy. It appears the attributes are
offset by one when they are written (or read) somehow. The cursor
position is also odd. The module is quite complex and difficult to
follow/modify, this should be a trivial fix (and probably is).
* Changed the 'signals' TCL script, it is slightly improved but still
seg-faults if I try to make it do what I want it to do, which is
automatically apply a signal filter.